requirements.txt
appdirs==1.4.4
astroid==2.3.3
beautifulsoup4==4.8.2
bottle==0.12.18
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
dateutils==0.6.11
docopt==0.6.2
extras==1.0.0
fake-useragent==0.1.11
fixtures==3.0.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
linecache2==1.0.0
lxml==4.5.2
mccabe==0.6.1
mock==4.0.2
num2words==0.5.10
parse==1.18.0
pbr==5.5.1
purl==1.5
pyee==7.0.4
pylint==2.4.4
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.8.1
python-dotenv==0.10.5
python-mimeparse==1.6.0
python-qbittorrent==0.4.2
pytz==2020.1
requests==2.22.0
requests-html==0.10.0
six==1.14.0
soupsieve==1.9.5
testscenarios==0.5.0
testtools==2.4.0
tqdm==4.50.2
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.25.7
w3lib==1.22.0
websockets==8.1
wrapt==1.11.2
